Revert "Versão 0.1"
This reverts commit ec10288abbde263b0cb17376f4a6a5695c31e25d.

+++ src/main/java/br/edu/ifpb/praticas/sparta/controladores/ControladorAtendimento.java
@@ -33,7 +33,7 @@ public class ControladorAtendimento implements Serializable{
     private Atendimento atend;
     
     public void agendamento() throws SQLException{
-        long cod = atend.getCodigo();
+        int cod = atend.getCodigo();
         int cliente = atend.getCliente();
         int tipo = atend.getTipoServico();
         Date data = atend.getData();
