+++ src/main/java/br/edu/ifpb/praticas/sparta/entidades/Pesquisa.java
@@ -15,26 +15,26 @@ public class Pesquisa implements Serializable{
 
     @Id
     @GeneratedValue
-    private long id;
+    private int id;
     @Column(nullable = false, unique = true)
-    private long atendimento;
+    private int atendimento;
     @Column(columnDefinition = "BOOLEAN DEFAULT FALSE")
     private boolean respondido;
     private float nota;
 
     public Pesquisa(){ }
 
-    public Pesquisa(long id, long atendimento, boolean respondido, float nota){
+    public Pesquisa(int id, int atendimento, boolean respondido, float nota){
         this.id = id;
         this.atendimento = atendimento;
         this.respondido = respondido;
         this.nota = nota;
     }
 
-    public long getId() {return id;}
-    public void setId(long id) {this.id = id;}
-    public long getAtendimento() {return atendimento;}
-    public void setAtendimento(long atendimento) {this.atendimento = atendimento;}
+    public int getId() {return id;}
+    public void setId(int id) {this.id = id;}
+    public int getAtendimento() {return atendimento;}
+    public void setAtendimento(int atendimento) {this.atendimento = atendimento;}
     public boolean isRespondido() {return respondido;}
     public void setRespondido(boolean respondido) {this.respondido = respondido;}
     public float getNota() {return nota;}
